Refactoring. Added wallfield and pathfield
