+++ towers/towers.py
@@ -1,0 +1,10 @@
+
+class Tower:
+    marker="T"
+    price=30
+
+class CannonTower(Tower):
+    marker="C"
+    price=30
+
+tower_dict = {"C": CannonTower}
