+++ waves/monsters.py
